20.6.Refs{Updating the repLogs State; Using uuids}
Each time we create a handler function in a class, we need to bind it to this! In the constructor add this.handleNewItemSubmit = the same thing .bind(this).

Using uuids

Try it again! We got it! It updates the state and that causes React to re-render and add the row. But... if we try it a second time, it does update the state, but, ah! It yells at us:

  Encountered two children with the same key

Ah! The id property is eventually used in RepLogList as the key prop. And with the hardcoded TODO, it's not unique. Time to fix that temporary hack.