20.7.Refs{Using uuids, yarn add uuid --dev}
But, hmmm. How can we get a unique id? There are always two options. First, you can make an AJAX request and wait for the server to send back the new id before updating the state. We'll do that later. Or, you can generate a uuid in JavaScript. Let's do that now. And later, when we start talking to the server via AJAX, we'll discuss how UUIDs can still be used, and are a great idea!

To generate a UUID, find your terminal and install a library:
yarn add uuid --dev