21.4.Immutability / Don't Mutate my State! {Moving the "itemOptions" onto a Property}
Moving the "itemOptions" onto a Property

Finally, we need to make one other small change. In RepLogCreator, all of our options are hardcoded. And, that's not necessarily a problem: we'll talk later about whether or not we should load these dynamically from the server.

But, to help show off some features we're about to work on, we need to make these a little bit more systematic. In the constructor, create a new property: this.itemOptions set to a data structure that represents the 4 items.

Notice, I'm not making this props or state: we don't need these options to actually change. Nope, we're just taking advantage of the fact that we have a class, so, if we want to, we can store some data on it.

Back in render(), delete the 4 options and replace it with one of our fancy map structures: this.itemOptions.map() with an item argument. In the function, return an <option> element with value={option.id}, key={option.id} - we need that for any array of elements - and, for the text, use {option.text}.

Nice! Let's make sure it works - refresh! It works and... yea - the options are still there.

Ok, try it again. Refresh, fill out the form and... we're good!

Let's talk about some validation!