22.2.Dumb Components with State{Preventing the Form Submit}
Preventing the Form Submit

The form handling-logic lives in RepLogCreator.handleFormSubmit(). So, if we want to prevent the form from submitting or add an error message, that code needs to live here.

And the first half is easy! If quantityInput.value <= 0 then... somehow, we need to add an error message. And because we don't want to continue processing, just return.

Let's see how this looks so far. Try negative 10, select an item and... yep! The form does not clear.