23.1.Form Validation State
When the user tries to submit a negative quantity, we need to show an error! On a React level, this means that we'll need new state to store that message. And, like we just discussed, because this state is really tied to our form and is meant to power the UI, the best place for it is inside RepLogCreator.

To add state, we always start the same way: initialize it in the constructor: this.state = an object, and call it quantityInputError set to empty quotes.

Below, remove the todo and add this.setState() with quantityInputError set to Please enter a value greater than zero.

And when the form submit is successful, we need to make sure any existing error is removed. Copy the setState() line and set it to an empty string.

State, check! Down in render, as always, start by destructuring the variable: const { quantityInputError } = this.state. Let's use this to do two things: add a class to make the form element look red and print the message.

Because we're using Bootstrap, to make the field red, the form-group div needs a new has-error class. Empty out className, enter into JavaScript mode and use "ticks" to use a "template literal". This makes using multiple classes with logic a bit easier: re-add form-group, then type ${} to do "string interpolation". Oooo. Inside, if quantityInputError then print has-error, else print nothing.