23.2.Form Validation State {Print the error message: The Cool Conditional String Printing Syntax}
The Cool Conditional String Printing Syntax

And to actually print the message, go down after the input. Here, I don't just want to print the error, I want to surround it in a red div, but only if there actually is an error. We could use the ternary syntax here... and use some inline JSX. That's fine.

But instead, I want to show you a cool, but weird, shortcut syntax that we can use whenever we want to print a string only when that string is not empty. Here it is: quantityInputError && and then the JSX: <span className="help-block">, print quantityInputError, close the tag, and exit from JavaScript mode.

Woh. Before we talk about this, try it! Move over, make sure the page is fully refreshed, select an item, be annoying and use a negative number and... there it is! Oh, it's ugly: we'll fix that soon.

But first, about this syntax! It's weird because this bit of code works different in JavaScript versus PHP! In JavaScript, if quantityInputError is empty, or "falsey", this return false and we print nothing. But if quantityInputError has some text in it, so, it's "truthy", then JavaScript returns the second part of the expression: our JSX. So, this entire block will either return false or this JSX. In PHP, this would always return false or true.

So... yes, this is another fancy syntax. If you love it, nice! If you hate it, use the ternary syntax instead.