24.4.Controlled Form Input{Casting to an Integer}
Casting to an Integer

Befire this code change we failed prop validation!

  numberOfHearts of type string supplied, expected number.

Interesting. In RepLogs, we expected the numberOfHearts prop to be a number... which makes sense. But apparently, it was a string! This wasn't that important... but it is interesting!

When you read a value from a field, it is, of course, always a string! That means the numberOfHearts state becomes a string and that is passed down as a prop. Let's clean that up: we could do that right here, or inside the handler function. To do it here, oh, this is bizarre, add a + before the variable.

That will change the string to a number. There are other ways to do this - JavaScript is weird - but this is one way.

Try it again: change the value and... no error!

Welcome to the world of "controlled components"! It feels really good... but it can be a bit more work. Don't worry: in a few minutes, we'll talk about when to use this strategy versus the original.

Oh, but to make this a little bit more fun, change this to input type="range".

Try it! Super-fun-heart-slider!

Next, let's refactor RepLogCreator to use controlled components. This will be the best way to see the difference between each approach.