26.2.Deleting Items{Adding & Calling the Delete Handler Function}
Adding & Calling the Delete Handler Function

In RepLogApp, create that new function: handleDeleteRepLog, which is a great name, because this component doesn't know and doesn't care that a link will be used to delete rep logs. Nope, it's all about the data. Give this an id argument so we know which rep log to delete. Be lazy and log a todo.

Next, because we have a new handler method, make sure to bind it to this.

And finally, pass this as a new prop: onDeleteRepLog={this.handleDeleteRepLog}.

Our work here is done. Now, move to RepLogs. First, at the bottom, add this to propTypes: onDeleteRepLog is PropTypes.func.isRequired.

Above in the function, destructure onDeleteRepLog, find RepLogList, and pass this again as a prop: onDeleteRepLog={onDeleteRepLog}.

Finally, move to RepLogList. Start the same: add the new prop to propTypes and destructure the variable.

Ultimately, we need to execute this callback onClick() of the link. We have a choice here: create an inline arrow function, or add a function above render. If the logic is simple, both are fine. Add a new handleDeleteClick function with two arguments: the event and repLogId. Start with event.preventDefault() so the browser doesn't try to follow the link. Then, yep, just onDeleteRepLog(repLogId).

Scroll down to hook this up: onClick={}. Hmm, we can't call handleDeleteClick directly... because we also need to pass it the id. No worries: use an arrow function with (event) => handleDeleteClick() passing it event and - because we're inside the loop, repLog.id.

Let's try it! Refresh! It looks good... and click delete. Nothing happens, but check the console. Got it! There is our todo.