27.1.API Setup & AJAX with fetch() {The API Setup}
Ok people: it's time to make our React app legit, by loading and saving data to the server via AJAX. Our Symfony app already has a functional set of API endpoints to load, delete and save rep logs. We're not going to spend a lot of time talking about the API side of things: we'll save that for a future tutorial. That's when we'll also talk about other great tools - like ApiPlatform - that you won't see used here.

Anyways, I want to at least take you through the basics of my simple, but very functional, setup.

The API Setup

Open src/Controller/RepLogController.php. As you can see, we already have API endpoints for returning all of the rep logs for a user, a single rep log, deleting a rep log and adding a new rep log. Go back to the browser to check this out: /reps. Boom! A JSON list of the rep logs.

This endpoint is powered by getRepLogsAction(). The findAllUsersRepLogModels() method lives in the parent class - BaseController, which lives in this same directory. Hold Command or Ctrl and click to jump into it.

The really important part is this: I have two rep log classes. First, the RepLog entity stores all the data in the database. Second, in the Api directory, I have another class called RepLogApiModel. This is the class that's transformed into JSON and used for the API: you can see that it has the same fields as the JSON response.

The findAllUsersRepLogModels method first queries for the RepLog entity objects. Then, it loops over each and transforms it into a RepLogApiModel object by calling another method, which lives right above this. The code is super boring and not fancy at all: it simply takes the RepLog entity object and, piece by piece, converts it into a RepLogApiModel.

Finally, back in getRepLogsAction(), we return $this->createApiResponse() and pass it that array of RepLogApiModel objects. This method also lives inside BaseController and it's dead-simple: it uses Symfony's serializer to turn the objects to JSON, then puts that into a Response.

That's it! The most interesting part is that I'm using two classes: the entity and a separate class for the serializer. Having 2 classes means that you need to do some extra work. However... it makes it really easy to make your API look exactly how you want! But, in a lot of cases, serializing your entity object directly works great.