27.2.API Setup & AJAX with fetch() {Using fetch()}
Using fetch()

So here's our first goal: make an API request to /reps and use that to populate our initial repLogs state so that they render in the table.

In the assets/js directory, create a new folder called api and then a new file called rep_log_api.js. This new file will contain all of the logic we need for making requests related to the rep log API endpoints. As our app grows, we might create other files to talk to other resources, like "users" or "products".

You probably also noticed that the filename is lowercase. That's a minor detail. This is because, instead of exporting a class, this module will export some functions... so that's just a naming convention. Inside, export function getRepLogs().

The question now is... how do we make AJAX calls? There are several great libraries that can help with this. But... actually... we don't need them! All modern browsers have a built-in function that makes AJAX calls super easy. It's called fetch()!

Try this: return fetch('/reps'). fetch() returns a Promise object, which is a super important, but kinda-confusing object we talked a lot about in our ES6 tutorial. To decode the JSON from our API into a JavaScript object, we can add a success handler: .then(), passing it an arrow function with a response argument. Inside, return response.json().

With this code, our getRepLogs() function will still return a Promise. But the "data" for that should now be the decoded JSON. Don't worry, we'll show this in action.

By the way, I mentioned that fetch is available in all modern browsers. So yes, we do need to worry about what happens in older browsers. We'll do that later.

Go back to RepLogApp. Ok, as soon as the page loads, we want to make an AJAX call to /reps and use that to populate the state. The constructor seems like a good place for that code. Oh, but first, bring it in: import { getRepLogs } from ../api/rep_log_api. For the first time, we're not exporting a default value: we're exporting a named function. We'll export more named functions later, for inserting and deleting rep logs.

Oh, and, did you see how PhpStorm auto-completed that for me? That was awesome! And it wasn't video magic: PhpStorm was cool enough to guess that correct import path.

Down below, add getRepLogs() and chain .then(). Because we decoded the JSON already, this should receive that decoded data. Just log it for now.

Ok... let's try it! Move over and, refresh! Oof! An error:

  Unexpected token in JSON at position zero

Hmm. It seems like our AJAX call might be working... but it's having a problem decoding the JSON. It turns out, the problem is authentication. Let's learn how to debug this and how to authenticate our React API requests next.